feat(persist): abstract sqlite persister implementation
